nova: remove default and obsolete flags from nova.conf
Setting a flag to a default value in nova.conf isn't useful. It doesn't
even serve as useful documentation since only a tiny subset of the flags
are included.

Also, a number of the flags are now obsolete showing the difficulty in
maintaining such an approach.

Interesting to note that the only flags now set are paths relating to
how nova is packaged in Fedora.