nova: add workaround for sphinx-build segfault
https://bugzilla.redhat.com/707199

sphinx-build recently began segfaulting in koji when the nova_autodoc
plugin is auto-generating files.

It wasn't happening with the d4~20110816.1449 snapshot, but it's not
clear what changed since then to cause the issue. It may simply that the
size of stdout from the generate_autodoc_index.sh script has increased.

If we run generate_autodoc_index.sh directly before the build and
disable the nova_autodoc extension using $SPHINX_DEBUG, it seems to work
fine.

Note: in theory, we shouldn't have to disable the autodoc extension and
it wouldn't re-generate the files, but in order for that to work we'd
need this small fix:

   diff --git a/doc/find_autodoc_modules.sh b/doc/find_autodoc_modules.sh
   index c612b00..5bc8b46 100755
   --- a/doc/find_autodoc_modules.sh
   +++ b/doc/find_autodoc_modules.sh
   @@ -1,7 +1,7 @@
    #!/bin/bash

    NOVA_DIR='../nova/' # include trailing slash
   -DOCS_DIR='source'
   +DOCS_DIR='source/api'

    modules=''
    for x in `find ${NOVA_DIR} -name '*.py'`; do